Add description to test cases and add more test cases.
