+++ src/main/java/seedu/address/logic/commands/room/FindEmptyRoomCommand.java
@@ -13,13 +13,13 @@ import seedu.address.model.room.Room;
  * Finds the next empty room with the lowest room number
  */
 //@@author itssodium
-public class FindRoomCommand extends Command {
+public class FindEmptyRoomCommand extends Command {
     public static final String COMMAND_WORD = "findemptyroom";
     public static final String MESSAGE_NO_EMPTY_ROOM = "There are unfortunately no more empty rooms!";
     public static final String NUMBER_OF_ROOMS_UNDEFINED = "Please define number of rooms";
     public static final String MESSAGE_SUCCESS = "Room Number %s is empty";
 
-    public FindRoomCommand() {}
+    public FindEmptyRoomCommand() {}
 
     @Override
     public CommandResult execute(Model model) throws CommandException {
